L13/src/main/java/ru/otus/spring/petrova/service/AuthorService.java
package ru.otus.spring.petrova.service;

import lombok.RequiredArgsConstructor;
import org.hibernate.exception.ConstraintViolationException;
import org.springframework.stereotype.Service;
import ru.otus.spring.petrova.repository.AuthorRepository;
import ru.otus.spring.petrova.domain.Author;
import ru.otus.spring.petrova.exception.AlreadyExist;

@Service
@RequiredArgsConstructor
public class AuthorService {

  private final AuthorRepository authorRepository;

  public void createAuthor(String name) throws AlreadyExist {
    try {
      authorRepository.save(new Author(name));
    } catch (RuntimeException e) {
      if (e.getCause() instanceof ConstraintViolationException) {
        throw new AlreadyExist("author", e);
      }
      throw e;
    }
  }
}
